requirements.txt
boto3==1.17.99
confluent_kafka==2.5.3
django>=3.2,<4.0
protobuf==5.28.0
pybase62
pynamodb==5.1.0
pynamodb-attributes==0.5.0
snowflake-id
PyJWT==2.1.0
pytest-django==4.4.0
pytest-mock==3.6.1